+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.4.1 (2022-03-30)
+------------------
+
+* Feature: enable package extras to be given in development mode
+* Feature: add support for arm64/aarch64-based development by forcing experiment containers to run as amd64 containers on those systems (#32)
+* Fix: add gym to extended package requirements
+
+
 0.4.0 (2022-03-28)
 ------------------